Implement TWTS-laser twtstight to replace twtsfast
